Reworked easy mode + rules description
Easy mode was too easy for the most, so now you only have 10 lives for every round to balance the situation. other then that i chanced the description for on the start menu, every times you hover over a button it tells you what it does. small stylesheet changes and new sound